+++ packages/core/rslib.config.ts
@@ -27,11 +27,16 @@ export default defineConfig({
 	lib: [
 		{
 			bundle: false,
-			dts: true,
+			dts: {
+				autoExtension: true,
+			},
 			format: "esm",
 		},
 		{
 			bundle: false,
+			dts: {
+				autoExtension: true,
+			},
 			format: "cjs",
 		},
 	],